project/requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Bcrypt>=1.0.1
Flask-Login>=0.6.3
pymysql>=1.1.0
pandas>=2.1.3
textblob>=0.17.1
nltk>=3.8.1
openpyxl>=3.1.2
xlrd>=2.0.1
reportlab>=4.0.7
python-dotenv>=1.0.0
Werkzeug>=3.0.1
transformers>=4.35.0
torch>=2.1.0
scipy>=1.11.0
requests>=2.1.0
Flask-WTF>=1.2.0
Flask-Limiter>=3.5.0
scikit-learn
cachetools
itsdangerous>=2.1.0
gunicorn>=21.2.0
flask-migrate>=4.0.0
pyotp>=2.9.0
qrcode[pil]>=7.4.0
psycopg2-binary>=2.9.0
